# static/swagger.yml
openapi: 3.0.0
paths:
...
  /api/v2/user:
    post:
      operationId: users-get
      summary: Получение информации о пользователях (только администратор)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsersGetRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersGetResponse"
        "400": &ref_54
          description: Ответ будет таким если с данным что-то не так
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401": &ref_55
          description: Ответ для незарегистрированного пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403": &ref_56
          description: Ответ для неавторизованного пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404": &ref_57
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "406": &ref_58
          description: Не принято значение
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
        "409": &ref_59
          description: Ответ для конфликта файлов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "412": &ref_60
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionFailedError"
        "500": &ref_61
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
        "501": &ref_62
          description: Пока не реализовано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
      tags: &ref_63
        - user
      security: &ref_64
        - bearer: []
  /api/v2/user/disable/{userId}:
    patch:
      operationId: user-disable
      summary: Скрытие аккаунта пользователя (только администратор)
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": *ref_54
        "401": *ref_55
        "403": *ref_56
        "404": *ref_57
        "406": *ref_58
        "409": *ref_59
        "412": *ref_60
        "500": *ref_61
        "501": *ref_62
      tags: *ref_63
      security: *ref_64
  /api/v2/user/enable/{userId}:
    patch:
      operationId: user-enable
      summary: Открытие аккаунта пользователя (только администратор)
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": *ref_54
        "401": *ref_55
        "403": *ref_56
        "404": *ref_57
        "406": *ref_58
        "409": *ref_59
        "412": *ref_60
        "500": *ref_61
        "501": *ref_62
      tags: *ref_63
      security: *ref_64
  /api/v2/user/{userId}:
    get:
      operationId: user-get
      summary: Получение информации о пользователе (только администратор)
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGetResponse"
        "400": *ref_54
        "401": *ref_55
        "403": *ref_56
        "404": *ref_57
        "406": *ref_58
        "409": *ref_59
        "412": *ref_60
        "500": *ref_61
        "501": *ref_62
      tags: *ref_63
      security: *ref_64
    patch:
      operationId: user-update
      summary: Изменение информации о пользователе (только администратор)
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGetResponse"
        "400": *ref_54
        "401": *ref_55
        "403": *ref_56
        "404": *ref_57
        "406": *ref_58
        "409": *ref_59
        "412": *ref_60
        "500": *ref_61
        "501": *ref_62
      tags: *ref_63
      security: *ref_64
    delete:
      operationId: user-delete
      summary: Удаление аккаунта пользователя (только администратор)
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": *ref_54
        "401": *ref_55
        "403": *ref_56
        "404": *ref_57
        "406": *ref_58
        "409": *ref_59
        "412": *ref_60
        "500": *ref_61
        "501": *ref_62
      tags: *ref_63
      security: *ref_64
...
info:
  title: MyScreen backend API
  description: MyScreen backend API
  version: 2.41.29
  contact:
    name: Stanislav V Vyaliy
    url: https://wisekaa.dev
    email: stanislav@wisekaa.dev
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Токен авторизации
      name: token
  schemas:
    BadRequestError:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        code:
          type: string
          example: server-error.10004
        message:
          type: string
          example: Bad request
      required:
        - statusCode
        - code
        - message
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 401
        code:
          type: string
          example: server-error.10001
        message:
          type: string
          example: Unauthorized request
      required:
        - statusCode
        - code
        - message
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: number
          example: 403
        code:
          type: string
          example: server-error.10002
        message:
          type: string
          example: Forbidden
      required:
        - statusCode
        - code
        - message
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: number
          example: 404
        code:
          type: string
          example: server-error.10005
        message:
          type: string
          example: Not Found
      required:
        - statusCode
        - code
        - message
    NotAcceptableError:
      type: object
      properties:
        statusCode:
          type: number
          example: 406
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Not Acceptable
      required:
        - statusCode
        - code
        - message
    FileIDResponse:
      type: object
      properties:
        id:
          type: string
          description: Идентификатор файла
          format: uuid
        name:
          type: string
          description: Имя файла
          example: foo.mp4
      required:
        - id
        - name
    ConflictDataFile:
      type: object
      properties:
        id:
          type: string
          description: ID ссылки
          format: uuid
        name:
          type: string
          description: Имя ссылки
          example: Имя ссылки
        file:
          description: Файл
          allOf:
            - $ref: "#/components/schemas/FileIDResponse"
    ConflictData:
      type: object
      properties:
        video:
          description: Редакторы (видео)
          allOf:
            - $ref: "#/components/schemas/ConflictDataFile"
        audio:
          description: Редакторы (аудио)
          allOf:
            - $ref: "#/components/schemas/ConflictDataFile"
        playlist:
          description: Плэйлисты
          allOf:
            - $ref: "#/components/schemas/ConflictDataFile"
        monitor:
          description: Мониторы
          allOf:
            - $ref: "#/components/schemas/ConflictDataFile"
    ConflictError:
      type: object
      properties:
        statusCode:
          type: number
          example: 409
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Conflict exists
        data:
          $ref: "#/components/schemas/ConflictData"
      required:
        - statusCode
        - code
        - message
        - data
    PreconditionFailedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 412
        code:
          type: string
          example: server-error.10002
        message:
          type: string
          example: User exists
      required:
        - statusCode
        - code
        - message
    InternalServerError:
      type: object
      properties:
        statusCode:
          type: number
          example: 500
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Server error
      required:
        - statusCode
        - code
        - message
    NotImplementedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 501
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Server error
      required:
        - statusCode
        - code
        - message
    Status:
      type: string
      enum:
        - success
        - error
    UserRoleResponse:
      type: string
      enum:
        - monitor
        - monitor-owner
        - advertiser
        - administrator
        - accountant
    UserPlan:
      type: string
      enum:
        - DEMO
        - FULL
        - VIP
    UserWallet:
      type: object
      properties:
        total:
          type: number
          description: Баланс
          example: 0
    UserMetricsMonitors:
      type: object
      properties:
        online:
          type: number
          description: Кол-во устройств с заявками на трансляции
        offline:
          type: number
          description: Кол-во устройств с заявками, но выключенные
        empty:
          type: number
          description: Кол-во устройств без заявок
        user:
          type: number
          description: Кол-во моих мониторов
      required:
        - online
        - offline
        - empty
        - user
    UserMetricsStorage:
      type: object
      properties:
        storage:
          type: number
          description: Занятое место
        total:
          type: number
          description: Максимальное место
      required:
        - storage
        - total
    UserMetricsPlaylists:
      type: object
      properties:
        added:
          type: number
          description: Добавленные
        played:
          type: number
          description: Запущенные
      required:
        - added
        - played
    UserMetrics:
      type: object
      properties:
        monitors:
          description: Статистика мониторов
          allOf:
            - $ref: "#/components/schemas/UserMetricsMonitors"
        storageSpace:
          description: Дисковое пространство
          allOf:
            - $ref: "#/components/schemas/UserMetricsStorage"
        playlists:
          description: Плейлисты
          allOf:
            - $ref: "#/components/schemas/UserMetricsPlaylists"
      required:
        - monitors
        - storageSpace
        - playlists
    UserExtEntity:
      type: object
      properties:
        id:
          type: string
          description: Идентификатор пользователя
          format: uuid
        email:
          type: string
          description: EMail пользователя
          format: email
          example: foo@bar.baz
        surname:
          type: string
          description: Фамилия
          example: Steve
          nullable: true
        name:
          type: string
          description: Имя
          example: John
          nullable: true
        middleName:
          type: string
          description: Отчество
          example: Doe
          nullable: true
        phoneNumber:
          type: string
          description: Телефон пользователя
          example: "+78002000000"
          nullable: true
        city:
          type: string
          description: Город
          example: Krasnodar
          nullable: true
        country:
          type: string
          description: Страна
          example: RU
          nullable: true
        storageSpace:
          type: number
          description: Дисковое пространство
          example: 20000000
        role:
          description: Роль пользователя
          example: advertiser
          $ref: "#/components/schemas/UserRoleResponse"
        verified:
          type: boolean
          description: EMail подтвержден
          example: true
        plan:
          description: План пользователя
          example: FULL
          $ref: "#/components/schemas/UserPlan"
        company:
          type: string
          description: Компания
          example: ACME corporation
          nullable: true
        companyLegalAddress:
          type: string
          description: Юридический адрес
          example: г. Краснодар, ул. Красная, д. 1
        companyActualAddress:
          type: string
          description: Фактический адрес
          example: г. Краснодар, ул. Красная, д. 1
        companyTIN:
          type: string
          description: Идентификационный номер налогоплательщика (ИНН)
          example: "012345678901"
        companyRRC:
          type: string
          description: Код Причины Постановки на учет (КПП)
          example: "012345678901"
        companyPSRN:
          type: string
          description: Основной Государственный Регистрационный Номер (ОГРН)
          example: "012345678901"
        companyPhone:
          type: string
          description: Телефон организации
          example: "+78002000000"
        companyEmail:
          type: string
          description: Email организации
          example: we@are.the.best
        companyBank:
          type: string
          description: Наименование банка
          example: Банк
        companyBIC:
          type: string
          description: Банковский идентификационный код (БИК)
          example: "012345678"
        companyCorrespondentAccount:
          type: string
          description: Корреспондентский счет
          example: "30101810400000000000"
        companyPaymentAccount:
          type: string
          description: Расчетный счет
          example: "40802810064580000000"
        companyFax:
          type: string
          description: Факс организации
          example: "+78002000000"
        companyRepresentative:
          type: string
          description: Представитель организации
          example: Тухбатуллина Юлия Евгеньевна
        createdAt:
          format: date-time
          type: string
          description: Время создания
          example: "2021-01-01T10:00:00.147Z"
        updatedAt:
          format: date-time
          type: string
          description: Время изменения
          example: "2021-01-01T10:00:00.147Z"
        planValidityPeriod:
          type: number
          description: Оставшийся срок оплаты
        wallet:
          description: Баланс
          allOf:
            - $ref: "#/components/schemas/UserWallet"
        metrics:
          description: Метрика
          allOf:
            - $ref: "#/components/schemas/UserMetrics"
      required:
        - id
        - email
        - role
        - createdAt
        - updatedAt
    UserGetResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
        data:
          description: Пользователь
          allOf:
            - $ref: "#/components/schemas/UserExtEntity"
      required:
        - status
        - data
    UserRoleRequest:
      type: string
      enum:
        - monitor-owner
        - advertiser
    UserUpdateRequest:
      type: object
      properties:
        email:
          type: string
          description: EMail пользователя
          format: email
          example: foo@bar.baz
        surname:
          type: string
          description: Фамилия
          example: Steve
          nullable: true
        name:
          type: string
          description: Имя
          example: John
          nullable: true
        middleName:
          type: string
          description: Отчество
          example: Doe
          nullable: true
        phoneNumber:
          type: string
          description: Телефон пользователя
          example: "+78002000000"
          nullable: true
        city:
          type: string
          description: Город
          example: Krasnodar
          nullable: true
        country:
          type: string
          description: Страна
          example: RU
          nullable: true
        verified:
          type: boolean
          description: EMail подтвержден
          example: true
        company:
          type: string
          description: Компания
          example: ACME corporation
          nullable: true
        companyLegalAddress:
          type: string
          description: Юридический адрес
          example: г. Краснодар, ул. Красная, д. 1
        companyActualAddress:
          type: string
          description: Фактический адрес
          example: г. Краснодар, ул. Красная, д. 1
        companyTIN:
          type: string
          description: Идентификационный номер налогоплательщика (ИНН)
          example: "012345678901"
        companyRRC:
          type: string
          description: Код Причины Постановки на учет (КПП)
          example: "012345678901"
        companyPSRN:
          type: string
          description: Основной Государственный Регистрационный Номер (ОГРН)
          example: "012345678901"
        companyPhone:
          type: string
          description: Телефон организации
          example: "+78002000000"
        companyEmail:
          type: string
          description: Email организации
          example: we@are.the.best
        companyBank:
          type: string
          description: Наименование банка
          example: Банк
        companyBIC:
          type: string
          description: Банковский идентификационный код (БИК)
          example: "012345678"
        companyCorrespondentAccount:
          type: string
          description: Корреспондентский счет
          example: "30101810400000000000"
        companyPaymentAccount:
          type: string
          description: Расчетный счет
          example: "40802810064580000000"
        companyFax:
          type: string
          description: Факс организации
          example: "+78002000000"
        companyRepresentative:
          type: string
          description: Представитель организации
          example: Тухбатуллина Юлия Евгеньевна
        role:
          description: Роль пользователя
          example: advertiser
          $ref: "#/components/schemas/UserRoleRequest"
        plan:
          description: План пользователя
          example: FULL
          $ref: "#/components/schemas/UserPlan"
...
    SuccessResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
      required:
        - status
...
    Order:
      type: string
      enum:
        - ASC
        - DESC
    LimitOrderRequest:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Order"
        name:
          $ref: "#/components/schemas/Order"
        description:
          $ref: "#/components/schemas/Order"
        extension:
          $ref: "#/components/schemas/Order"
        width:
          $ref: "#/components/schemas/Order"
        height:
          $ref: "#/components/schemas/Order"
        folder:
          $ref: "#/components/schemas/Order"
        folderId:
          $ref: "#/components/schemas/Order"
        monitor:
          $ref: "#/components/schemas/Order"
        monitorId:
          $ref: "#/components/schemas/Order"
        videoType:
          $ref: "#/components/schemas/Order"
        filesize:
          $ref: "#/components/schemas/Order"
        category:
          $ref: "#/components/schemas/Order"
        duration:
          $ref: "#/components/schemas/Order"
        files:
          $ref: "#/components/schemas/Order"
        monitors:
          $ref: "#/components/schemas/Order"
        playlists:
          $ref: "#/components/schemas/Order"
        parentFolder:
          $ref: "#/components/schemas/Order"
        parentFolderId:
          $ref: "#/components/schemas/Order"
        email:
          $ref: "#/components/schemas/Order"
        surname:
          $ref: "#/components/schemas/Order"
        middleName:
          $ref: "#/components/schemas/Order"
        phoneNumber:
          $ref: "#/components/schemas/Order"
        city:
          $ref: "#/components/schemas/Order"
        country:
          $ref: "#/components/schemas/Order"
        company:
          $ref: "#/components/schemas/Order"
        role:
          $ref: "#/components/schemas/Order"
        verified:
          $ref: "#/components/schemas/Order"
        isDemoUser:
          $ref: "#/components/schemas/Order"
        countUsedSpace:
          $ref: "#/components/schemas/Order"
        address:
          $ref: "#/components/schemas/Order"
        price:
          $ref: "#/components/schemas/Order"
        orientation:
          $ref: "#/components/schemas/Order"
        monitorInfo:
          $ref: "#/components/schemas/Order"
        attached:
          $ref: "#/components/schemas/Order"
        code:
          $ref: "#/components/schemas/Order"
        status:
          $ref: "#/components/schemas/Order"
        longitude:
          $ref: "#/components/schemas/Order"
        latitude:
          $ref: "#/components/schemas/Order"
        renderingStatus:
          $ref: "#/components/schemas/Order"
        fps:
          $ref: "#/components/schemas/Order"
        renderingPercent:
          $ref: "#/components/schemas/Order"
        renderingError:
          $ref: "#/components/schemas/Order"
        renderedFile:
          $ref: "#/components/schemas/Order"
        renderedFileId:
          $ref: "#/components/schemas/Order"
        keepSourceAudio:
          $ref: "#/components/schemas/Order"
        totalDuration:
          $ref: "#/components/schemas/Order"
        videoLayers:
          $ref: "#/components/schemas/Order"
        audioLayers:
          $ref: "#/components/schemas/Order"
        createdAt:
          $ref: "#/components/schemas/Order"
        updatedAt:
          $ref: "#/components/schemas/Order"
        seller:
          $ref: "#/components/schemas/Order"
        buyer:
          $ref: "#/components/schemas/Order"
        playlist:
          $ref: "#/components/schemas/Order"
        favorite:
          $ref: "#/components/schemas/Order"
    LimitRequest:
      type: object
      properties:
        limit:
          type: number
          description: Лимит строк результатов
          example: 20
        page:
          type: number
          description: Страница результатов
          example: 1
        order:
          description: Порядок результатов
          example:
            createdAt: DESC
          allOf:
            - $ref: "#/components/schemas/LimitOrderRequest"
...
    UserRequest:
      type: object
      properties:
        id:
          type: string
          description: Идентификатор пользователя
          format: uuid
        email:
          type: string
          description: EMail пользователя
          format: email
          example: foo@bar.baz
        surname:
          type: string
          description: Фамилия
          example: Steve
          nullable: true
        name:
          type: string
          description: Имя
          example: John
          nullable: true
        middleName:
          type: string
          description: Отчество
          example: Doe
          nullable: true
        password:
          type: string
          example: Secret~12345678
          description: Пароля пользователя (должен удовлетворять минимальным требованиям)
          minLength: 8
          maxLength: 30
          pattern: |-
            /((?=.*d)|(?=.*W+))(?![.
            ])(?=.*[A-Z])(?=.*[a-z]).*$/
        phoneNumber:
          type: string
          description: Телефон пользователя
          example: "+78002000000"
          nullable: true
        city:
          type: string
          description: Город
          example: Krasnodar
          nullable: true
        country:
          type: string
          description: Страна
          example: RU
          nullable: true
        storageSpace:
          type: number
          description: Дисковое пространство
          example: 20000000
        role:
          description: Роль пользователя
          example: advertiser
          $ref: "#/components/schemas/UserRoleResponse"
        verified:
          type: boolean
          description: EMail подтвержден
          example: true
        plan:
          description: План пользователя
          example: FULL
          $ref: "#/components/schemas/UserPlan"
        company:
          type: string
          description: Компания
          example: ACME corporation
          nullable: true
        companyLegalAddress:
          type: string
          description: Юридический адрес
          example: г. Краснодар, ул. Красная, д. 1
        companyActualAddress:
          type: string
          description: Фактический адрес
          example: г. Краснодар, ул. Красная, д. 1
        companyTIN:
          type: string
          description: Идентификационный номер налогоплательщика (ИНН)
          example: "012345678901"
        companyRRC:
          type: string
          description: Код Причины Постановки на учет (КПП)
          example: "012345678901"
        companyPSRN:
          type: string
          description: Основной Государственный Регистрационный Номер (ОГРН)
          example: "012345678901"
        companyPhone:
          type: string
          description: Телефон организации
          example: "+78002000000"
        companyEmail:
          type: string
          description: Email организации
          example: we@are.the.best
        companyBank:
          type: string
          description: Наименование банка
          example: Банк
        companyBIC:
          type: string
          description: Банковский идентификационный код (БИК)
          example: "012345678"
        companyCorrespondentAccount:
          type: string
          description: Корреспондентский счет
          example: "30101810400000000000"
        companyPaymentAccount:
          type: string
          description: Расчетный счет
          example: "40802810064580000000"
        companyFax:
          type: string
          description: Факс организации
          example: "+78002000000"
        companyRepresentative:
          type: string
          description: Представитель организации
          example: Тухбатуллина Юлия Евгеньевна
        createdAt:
          description: Время создания
          example:
            - "2021-01-01"
            - "2021-12-31"
          type: array
          items:
            type: array
        updatedAt:
          description: Время изменения
          example:
            - "2021-01-01"
            - "2021-12-31"
          type: array
          items:
            type: array
    UsersGetRequest:
      type: object
      properties:
        where:
          description: Запрос
          allOf:
            - $ref: "#/components/schemas/UserRequest"
        select:
          type: array
          description: Выбрать поля
          example: []
          items:
            type: string
            enum:
              - id
              - email
              - surname
              - name
              - middleName
              - phoneNumber
              - city
              - country
              - storageSpace
              - role
              - verified
              - plan
              - company
              - companyLegalAddress
              - companyActualAddress
              - companyTIN
              - companyRRC
              - companyPSRN
              - companyPhone
              - companyEmail
              - companyBank
              - companyBIC
              - companyCorrespondentAccount
              - companyPaymentAccount
              - companyFax
              - companyRepresentative
              - createdAt
              - updatedAt
        scope:
          description: Рамки для запроса
          allOf:
            - $ref: "#/components/schemas/LimitRequest"
    UsersGetResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
        count:
          type: number
          description: Количество пользователей
        data:
          description: Пользователи
          type: array
          items:
            $ref: "#/components/schemas/UserExtEntity"
      required:
        - status
        - count
        - data
...